+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 - Added automatic retry for validate session at 30 second intervals if it fails
 ### Changed
 - Now displays "N/A" in the links column if the item has no links
+- Stash tabs are now required to be selected before saving a profile
+- Added an asterisk to profile name field in profile dialog
 ### Fixed
 - Fixed a bug where some uniques were incorrectly priced as 6-links